refactor(settings): delete three single-consumer settings-view controllers
`LatexRecommendedSettingsController` kept VS Code config access behind a
three-method port whose only implementation was three
`vscode.workspace.getConfiguration()` calls in its one consumer; inline it
there, moving the `explorer.autoRevealExclude` merge/unmerge semantics
verbatim.

`SettingsAgentFileController` had no constructor and no fields — two pure
functions in a class, constructed once to call them; its methods move to
`SettingsAgentDirectoryController`, which already owns custom-agent directory
and file-path planning.

`SettingsAgentVisibilityController` had one method fed the same state object
the factory hands `SettingsAgentCatalogController`; `setAllAgentsEnabled`
moves there with its identical-list short-circuit intact.

Desktop `applyLatexSettings` was a silent no-op for a section the desktop
LaTeX tab never renders; it now answers `unsupported(...)`.
