+++ packages/extension/src/settingsView/handlers/latexSettingsHandlers.ts
@@ -8,7 +8,6 @@ import * as vscode from 'vscode';
 
 import { workspaceSM } from '@common/state';
 import { LatexToolingController } from '@controllers/settingsView/LatexToolingController';
-import { LatexRecommendedSettingsController } from '@controllers/settingsView/LatexRecommendedSettingsController';
 import { LatexConfigPersistenceController } from '@controllers/settingsView/LatexConfigPersistenceController';
 import { platform } from '@platform/platform';
 import { SETTINGS_VIEW_COMMANDS } from '@shared/ipc';
@@ -29,27 +28,108 @@ import {
   type SettingsHandlerContext,
 } from './SettingsHandlerContext';
 
+type LatexRecommendedSettingField = 'outDir' | 'autoRevealExclude';
+
+/** A recommended setting whose target value is a single scalar. */
+interface LatexRecommendedScalarSetting {
+  kind: 'scalar';
+  key: string;
+  value: string;
+  field: LatexRecommendedSettingField;
+}
+
+/**
+ * A recommended setting whose target value is an object merged key-by-key
+ * into whatever the user already has set (rather than overwritten wholesale).
+ */
+interface LatexRecommendedObjectSetting {
+  kind: 'object';
+  key: string;
+  value: Record<string, unknown>;
+  field: LatexRecommendedSettingField;
+}
+
+type LatexRecommendedSetting =
+  LatexRecommendedScalarSetting | LatexRecommendedObjectSetting;
+
+/** Recommended LaTeX-related VS Code settings and their target values. */
+const LATEX_RECOMMENDED_SETTINGS: LatexRecommendedSetting[] = [
+  {
+    kind: 'scalar',
+    key: 'latex-workshop.latex.outDir',
+    value: '%DIR%/build/',
+    field: 'outDir',
+  },
+  {
+    kind: 'object',
+    key: 'explorer.autoRevealExclude',
+    value: { '**/build/': true },
+    field: 'autoRevealExclude',
+  },
+];
+
+/**
+ * Whether a recommended setting is explicitly set to its recommended value.
+ * An object setting counts as set when it contains every recommended entry;
+ * the user's other keys are irrelevant.
+ */
+function isRecommendedValueSet(field: LatexRecommendedSettingField): boolean {
+  const setting = LATEX_RECOMMENDED_SETTINGS.find(
+    (candidate) => candidate.field === field,
+  );
+  if (!setting) return false;
+  const inspection = vscode.workspace.getConfiguration().inspect(setting.key);
+  const explicitlySet =
+    inspection?.globalValue !== undefined ||
+    inspection?.workspaceValue !== undefined ||
+    inspection?.workspaceFolderValue !== undefined;
+  if (!explicitlySet) return false;
+
+  const current = vscode.workspace.getConfiguration().get(setting.key);
+  if (setting.kind === 'scalar') {
+    return current === setting.value;
+  }
+
+  if (typeof current !== 'object' || current === null) return false;
+  const currentObject = current as Record<string, unknown>;
+  return Object.entries(setting.value).every(
+    ([key, value]) => currentObject[key] === value,
+  );
+}
+
+/**
+ * The value to write for one recommended setting. Object settings merge into
+ * (and, on reset, unmerge out of) the user's existing global value so their
+ * unrelated keys survive both directions.
+ */
+function resolveUpdateValue(
+  setting: LatexRecommendedSetting,
+  reset: boolean,
+): unknown {
+  if (setting.kind === 'scalar') {
+    return reset ? undefined : setting.value;
+  }
+
+  const globalValue = vscode.workspace
+    .getConfiguration()
+    .inspect<Record<string, unknown>>(setting.key)?.globalValue;
+  const remaining =
+    typeof globalValue === 'object' && globalValue !== null
+      ? { ...globalValue }
+      : {};
+
+  if (reset) {
+    for (const recommendedKey of Object.keys(setting.value)) {
+      delete remaining[recommendedKey];
+    }
+    return Object.keys(remaining).length > 0 ? remaining : undefined;
+  }
+
+  return { ...remaining, ...setting.value };
+}
+
 /** LaTeX settings handler delegate. */
 export class LatexSettingsHandlers {
-  private readonly recommendedSettingsController =
-    new LatexRecommendedSettingsController({
-      config: {
-        getConfig: (key) => vscode.workspace.getConfiguration().get(key),
-        getGlobalValue: (key) =>
-          vscode.workspace
-            .getConfiguration()
-            .inspect<Record<string, unknown>>(key)?.globalValue,
-        isExplicitlySet: (key) => {
-          const inspection = vscode.workspace.getConfiguration().inspect(key);
-          return (
-            inspection?.globalValue !== undefined ||
-            inspection?.workspaceValue !== undefined ||
-            inspection?.workspaceFolderValue !== undefined
-          );
-        },
-      },
-    });
-
   private readonly toolingController = new LatexToolingController({
     checkToolInstalled: (tool) => checkToolInstalled(tool, false),
     findPath: (tool) => BinaryResolver.findPath(tool),
@@ -58,12 +138,8 @@ export class LatexSettingsHandlers {
     isLatexWorkshopInstalled: () =>
       Boolean(vscode.extensions.getExtension(LATEX_WORKSHOP_EXT_ID)),
     getRecommendedStatus: () => ({
-      outDir:
-        this.recommendedSettingsController.isRecommendedValueSet('outDir'),
-      autoRevealExclude:
-        this.recommendedSettingsController.isRecommendedValueSet(
-          'autoRevealExclude',
-        ),
+      outDir: isRecommendedValueSet('outDir'),
+      autoRevealExclude: isRecommendedValueSet('autoRevealExclude'),
     }),
     onDetectionError: (error) => {
       this.ctx.logger.error(
@@ -92,16 +168,20 @@ export class LatexSettingsHandlers {
       this.ctx,
       'Failed to update LaTeX settings',
       async () => {
-        for (const {
-          key,
-          value,
-        } of this.recommendedSettingsController.buildUpdates({
-          field: data.field,
-          reset: data.reset ?? false,
-        })) {
+        const reset = data.reset ?? false;
+        const targets = data.field
+          ? LATEX_RECOMMENDED_SETTINGS.filter(
+              (setting) => setting.field === data.field,
+            )
+          : LATEX_RECOMMENDED_SETTINGS;
+        for (const setting of targets) {
           await vscode.workspace
             .getConfiguration()
-            .update(key, value, vscode.ConfigurationTarget.Global);
+            .update(
+              setting.key,
+              resolveUpdateValue(setting, reset),
+              vscode.ConfigurationTarget.Global,
+            );
         }
 
         await this.ctx.withActiveWebview((w) =>
